Move common code out of `handlers.dart` to fix internal build.
