+++ packages/devtools_server/lib/devtools_server.dart
@@ -2,5 +2,5 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'src/handlers.dart';
+export 'src/external_handlers.dart';
 export 'src/server.dart';
